requirements.txt
paddleocr
easyocr
mistralai
pydantic
opencv-python
pdf2image
requests
pillow
python-dotenv
fastapi
uvicorn
httpx
playwright
pytesseract
pytest
python-multipart
paddlepaddle
PyJWT
cryptography